requirements.txt
fastapi
uvicorn[standard]
motor
redis[asyncio]
python-jose[cryptography]
httpx
pydantic
python-dotenv
